+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Version 0.6.1
 * Deep gets on nodes with keywords returns an enum like the regular get.
 * Fix rare failures of `wait_for_state_change` function that resulted in early timeouts.
+* All keywords on nodes are correctly supported, and not only the first one for each option.
 
 ## Version 0.6.0
 * Revert full support of `fnmatch` wildcards and instead use the LabOne wildcard support.